backend/httpServer/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
fake-useragent==2.0.3
groq==0.16.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
python-dotenv==1.0.1
Flask==3.1.0
gunicorn==21.2.0
openai==1.63.0
# flask-cors==5.0.0 # nginx will handle CORS